split shared utils module
The shared utils module was never a nice solution. But recently has
become grown even bigger. Too big. Looking closer it could be identified
that there are two groups of functions: those who do something with
buffers and others to work with actions. Therefore I split them like
that. I'm still not too happy with having these "utility" stuff here.
